Initial commit of Car City
I properly pushed what I think is the latest version of your game to a git repo.  Just in case I'm putting the wrong project in the repo, I put the zip you mentioned in your google drive. If I uploaded the correct project then you should delete the .zip and continue to work out of this repository using git.